requirements/base.txt
Django==2.1.5
django-crispy-forms==1.7.2
django-environ==0.4.5
django-rest-framework==0.1.0
djangorestframework==3.9.1
mysqlclient==1.4.1
